Added packing feature Supported DSAR Compression in Ghost of Tsushima (#6) Fixed some bugs and removed unused code

+++ Helpers/IOHelper.cs
@@ -1,30 +1,28 @@
 ﻿using Gibbed.IO;
-using System;
 using System.IO;
-using System.Linq;
 
 namespace UnPSARC
 {
     public static class IOHelper
     {
-        public static byte[] ReadAtOffset(this Stream s, long Offset, int Size)
+        public static byte[] ReadAtOffset(this Stream s, long Offset, long Size)
         {
             long pos = s.Position;
             s.Seek(Offset, SeekOrigin.Begin);
-            byte[] log = s.ReadBytes(Size);
+            byte[] log = s.ReadBytes((int)Size);
             s.Seek(pos, SeekOrigin.Begin);
             return log;
         }
-        public static byte[] ReadAtOffset(this Stream s, long Offset, int size, int ZSize, string CompressionType)
+        public static byte[] ReadAtOffset(this Stream s, long Offset, long size, int ZSize, string CompressionType)
         {
-            
+
             long pos = s.Position;
             s.Seek(Offset, SeekOrigin.Begin);
             int magic = s.ReadValueU16();
             s.Seek(Offset, SeekOrigin.Begin);
             byte[] Block = s.ReadBytes(ZSize);
             byte[] log = { };
-            if (CompressionType == "oodl" && magic == 0x68C) log = Oodle.Decompress(Block, size);
+            if (CompressionType == "oodl" && magic == 0x68C) log = Oodle.Decompress(Block, (int)size);
             if (CompressionType == "zlib" && magic == 0xDA78) log = Zlib.Decompress(Block);
             if (log.Length == 0) log = Block;
             s.Seek(pos, SeekOrigin.Begin);
